+++ src/NServiceBus.WindowsPerformanceCounters/Guard.cs
