+++ src/NServiceBus.WindowsPerformanceCounters/Metrics.cs
@@ -1,20 +1,5 @@
 ﻿namespace NServiceBus
 {
-    using Configuration.AdvanceExtensibility;
-
-    public static class MetricsExtensions
-    {
-        public static Metrics Metrics(this EndpointConfiguration endpointConfiguration)
-        {
-            Guard.AgainstNull(nameof(endpointConfiguration), endpointConfiguration);
-
-            // This is bogus, but perhaps usable to verify it we access the extensionmethod via Metrics()
-            endpointConfiguration.GetSettings().Set("Metrics", "Activated");
-
-            return new Metrics(endpointConfiguration);
-        }
-    }
-
     public class Metrics
     {
         internal Metrics(EndpointConfiguration endpointConfiguration)
