+++ src/NServiceBus.WindowsPerformanceCounters/PerformanceCounters.cs
